Restructured navigators order, so the tab bar does not get unnecesarely shown on screens that do not require it.
